+++ src/calculate/pass_4.rs
@@ -3,27 +3,27 @@ use super::{Info, Pass, Ship};
 
 pub struct PassFour {}
 
-fn add_attribute(ship: &mut Ship, attribute_id: i32, base_value: f32, value: f32) {
+fn add_attribute(ship: &mut Ship, attribute_id: i32, base_value: f64, value: f64) {
     let mut attribute = Attribute::new(base_value);
     attribute.value = Some(value);
     ship.hull.attributes.insert(attribute_id, attribute);
 }
 
-fn calculate_align_time(ship: &mut Ship) -> (f32, f32) {
+fn calculate_align_time(ship: &mut Ship) -> (f64, f64) {
     /* Align-time is based on agility (70) and mass (4). */
 
     let base_agility = ship.hull.attributes.get(&70).unwrap().base_value;
     let base_mass = ship.hull.attributes.get(&4).unwrap().base_value;
-    let base_align_time = -(0.25 as f32).ln() * base_agility * base_mass / 1000000.0;
+    let base_align_time = -(0.25 as f64).ln() * base_agility * base_mass / 1000000.0;
 
     let agility = ship.hull.attributes.get(&70).unwrap().value.unwrap();
     let mass = ship.hull.attributes.get(&4).unwrap().value.unwrap();
-    let align_time = -(0.25 as f32).ln() * agility * mass / 1000000.0;
+    let align_time = -(0.25 as f64).ln() * agility * mass / 1000000.0;
 
     (base_align_time, align_time)
 }
 
-fn add_scan_strength(ship: &mut Ship) -> (f32, f32) {
+fn add_scan_strength(ship: &mut Ship) -> (f64, f64) {
     /* Scan Strength can be one of 4 values. */
 
     let mut base_scan_strength = 0.0;
@@ -44,7 +44,7 @@ fn add_scan_strength(ship: &mut Ship) -> (f32, f32) {
     (base_scan_strength, scan_strength)
 }
 
-fn add_cpu_used(ship: &mut Ship) -> (f32, f32) {
+fn add_cpu_used(ship: &mut Ship) -> (f64, f64) {
     /* How much CPU is being used, which is adding up cpuOuput (50) from all items. */
 
     let mut cpu_used = 0.0;
@@ -57,7 +57,7 @@ fn add_cpu_used(ship: &mut Ship) -> (f32, f32) {
     (0.0, cpu_used)
 }
 
-fn add_pg_used(ship: &mut Ship) -> (f32, f32) {
+fn add_pg_used(ship: &mut Ship) -> (f64, f64) {
     /* How much PG is being used, which is adding up powerOutput (30) from all items. */
 
     let mut pg_used = 0.0;
@@ -70,7 +70,7 @@ fn add_pg_used(ship: &mut Ship) -> (f32, f32) {
     (0.0, pg_used)
 }
 
-fn add_cpu_unused(ship: &mut Ship) -> (f32, f32) {
+fn add_cpu_unused(ship: &mut Ship) -> (f64, f64) {
     /* How much CPU is left, which is the total CPU minus the usage. */
 
     let cpu_used = ship.hull.attributes.get(&-3).unwrap().value.unwrap();
@@ -80,7 +80,7 @@ fn add_cpu_unused(ship: &mut Ship) -> (f32, f32) {
     (0.0, cpu_unused)
 }
 
-fn add_pg_unused(ship: &mut Ship) -> (f32, f32) {
+fn add_pg_unused(ship: &mut Ship) -> (f64, f64) {
     /* How much PG is left, which is the total PG minus the usage. */
 
     let pg_used = ship.hull.attributes.get(&-4).unwrap().value.unwrap();
